+++ tdrive/frontend/src/app/views/client/body/drive/header-path.tsx
@@ -1,8 +1,11 @@
 import { Button } from '@atoms/button/button';
 import { Title } from '@atoms/text';
 import { DriveItem } from '@features/drive/types';
+import { ChevronDownIcon } from '@heroicons/react/solid';
 import { useEffect, useState } from 'react';
 import { PublicIcon } from './components/public-icon';
+import MenusManager from '@components/menus/menus-manager.jsx';
+import { useCurrentUser } from 'app/features/users/hooks/use-current-user';
 
 export default ({
   path: livePath,
@@ -56,6 +59,7 @@ const PathItem = ({
   first?: boolean;
   onClick: (id: string) => void;
 }) => {
+  const { user } = useCurrentUser();
   return (
     <Button
       theme={last ? 'primary' : 'default'}
@@ -65,8 +69,19 @@ const PathItem = ({
         (!last ? 'rounded-r-none ' : '') +
         (!first && !last ? 'max-w-[15ch] ' : '')
       }
-      onClick={() => {
-        onClick(item?.id || '');
+      onClick={evt => {
+        if (first) {
+          MenusManager.openMenu(
+            [
+              { type: 'menu', text: 'Home', onClick: () => onClick('root') },
+              { type: 'menu', text: 'My Drive', onClick: () => onClick('user_' + user?.id) },
+            ],
+            { x: evt.clientX, y: evt.clientY },
+            'center',
+          );
+        } else {
+          onClick(item?.id || '');
+        }
       }}
     >
       <span className="text-ellipsis overflow-hidden whitespace-nowrap" style={{ maxWidth: 120 }}>
@@ -74,6 +89,11 @@ const PathItem = ({
       </span>
       {item?.access_info?.public?.level && item?.access_info?.public?.level !== 'none' && (
         <PublicIcon className="h-5 w-5 ml-2" />
+      )}
+      {first && (
+        <span className="ml-2 -mr-1">
+          <ChevronDownIcon className="w-4 h-4" />
+        </span>
       )}
     </Button>
   );
