+++ tdrive/frontend/src/app/views/client/side-bar/index.tsx
@@ -6,7 +6,9 @@ import {
   HeartIcon,
   ShareIcon,
   TrashIcon,
+  UserIcon,
 } from '@heroicons/react/outline';
+import { useCurrentUser } from 'app/features/users/hooks/use-current-user';
 import { useRecoilState } from 'recoil';
 import { Title } from '../../../atoms/text';
 import { useDriveItem } from '../../../features/drive/hooks/use-drive-item';
@@ -18,10 +20,14 @@ import Actions from './actions';
 
 export default () => {
   const [parentId, setParentId] = useRecoilState(DriveCurrentFolderAtom('root'));
+  const { user } = useCurrentUser();
   const active = false;
   const { access: rootAccess } = useDriveItem('root');
-  const { inTrash } = useDriveItem(parentId);
+  const { inTrash, path } = useDriveItem(parentId);
   const activeClass = 'bg-zinc-50 dark:bg-zinc-800 !text-blue-500';
+  let folderType = 'home';
+  if ((path || [])[0]?.id === 'user_' + user?.id) folderType = 'personal';
+  if (inTrash) folderType = 'trash';
   return (
     <div className="grow flex flex-col overflow-auto -m-4 p-4 relative">
       <div className="grow">
@@ -45,9 +51,17 @@ export default () => {
           onClick={() => setParentId('root')}
           size="lg"
           theme="white"
-          className={'w-full mt-2 mb-1 ' + (!inTrash ? activeClass : '')}
+          className={'w-full mt-2 mb-1 ' + (folderType === 'home' ? activeClass : '')}
         >
           <CloudIcon className="w-5 h-5 mr-4" /> Home
+        </Button>
+        <Button
+          onClick={() => setParentId('user_' + user?.id)}
+          size="lg"
+          theme="white"
+          className={'w-full mb-1 ' + (folderType === 'personal' ? activeClass : '')}
+        >
+          <UserIcon className="w-5 h-5 mr-4" /> My Drive
         </Button>
         {false && (
           <>
@@ -72,7 +86,7 @@ export default () => {
             onClick={() => setParentId('trash')}
             size="lg"
             theme="white"
-            className={'w-full mb-1 ' + (inTrash ? activeClass : '')}
+            className={'w-full mb-1 ' + (folderType === 'trash' ? activeClass : '')}
           >
             <TrashIcon className="w-5 h-5 mr-4 text-rose-500" /> Trash
           </Button>
